requirements.txt
mysql.connector==2.2.9
email.mime.multipart==MIMEMultipart